[genio] link wrap _gettimeofday to fix broken mDNS
