+++ examples/platform/mt793x/BUILD.gn
@@ -30,8 +30,6 @@ config("chip_examples_project_config") {
     "-Wl,--wrap=calloc",
     "-Wl,--wrap=calloc",
     "-Wl,--wrap=MemoryAlloc",
-    "-Wl,--wrap=__assert_function",
-    "-Wl,--wrap=wlan_printf",
 
     # Wrap these in case internal newlib call them (e.g. strdup will)
     # directly call _malloc_r)
@@ -40,6 +38,10 @@ config("chip_examples_project_config") {
     "-Wl,--wrap=_free_r",
     "-Wl,--wrap=_calloc_r",
     "-Wl,--wrap=_write",
+    
+    "-Wl,--wrap=__assert_function",
+    "-Wl,--wrap=wlan_printf",
+    "-Wl,--wrap=_gettimeofday",
   ]
 }
 
@@ -48,11 +50,8 @@ source_set("genio-matter-shell") {
     defines = [ "ENABLE_CHIP_SHELL" ]
 
     sources = [
-      "assert_wrap.c",
-      "malloc_wrap.c",
       "matter_shell.cpp",
-      "wifi_wrap.c",
-      "wrap_write.c",
+      "link_wrapper.c",
     ]
     include_dirs = [ "." ]
