--- examples/platform/mt793x/wrap_write.c
@@ -1,8 +1,0 @@
-#include <stdio.h>
-
-extern int log_write(char * buf, int len);
-
-int __wrap__write(int file, char * ptr, int len)
-{
-    return log_write(ptr, len);
-}
